+++ docs/release_notes.rst
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+v4.2.1
+------
+
+* `SwiftPath.rmtree` on a container also deletes the segment container if it exists
+
 v4.2.0
 ------